refactor: MVC구조에 따라 controller 폴더 생성하여 racingGame.js파일 옮김
