+++ src/App.js
@@ -1,4 +1,4 @@
-import RacingGameController from "./racingGame.js";
+import RacingGameController from "./controller.js/racingGameController.js";
 class App {
   async run() {
     const gameController = new RacingGameController();
